Add user system.

+++ routes/articles.js
@@ -1,7 +1,9 @@
 const express = require("express");
 const _ = require("lodash");
 const Article = require("../models/article");
+const User = require("../models/user");
 const existenceVerifier = require("../helpers/existenceVerifier");
+const verifyToken = require("../middlewares/verifyToken");
 
 const router = express.Router();
 
@@ -79,9 +81,17 @@ router.get("/:id", (req, res) => {
  * @param {Article} req.body
  * @returns {String} Location header
  */
-// TODO: Add situation when same ID submit.
-router.post("/", async (req, res) => {
-  // req.body.createdBy = req.id;
+router.post("/", verifyToken, async (req, res) => {
+  // 只有管理员能够发布文章。
+  if (!(await existenceVerifier(User, { _id: req.id, group: "admin" }))) {
+    res.status(401).send("401 Unauthorized: Insufficient permissions.");
+    return;
+  }
+
+  req.body.createdBy = req.id;
+  req.createdAt = new Date().toISOString();
+  req.body.updatedBy = req.id;
+  req.updatedAt = new Date().toISOString();
   const newArticle = new Article(req.body);
   newArticle.save((err, article) => {
     if (err) {
@@ -99,7 +109,13 @@ router.post("/", async (req, res) => {
  * @param {String} id 需要更新的文章 ID
  * @returns {String} Location header 或 空
  */
-router.put("/:id", async (req, res) => {
+router.put("/:id", verifyToken, async (req, res) => {
+  // 只有管理员能够更新文章。
+  if (!(await existenceVerifier(User, { _id: req.id, group: "admin" }))) {
+    res.status(401).send("401 Unauthorized: Insufficient permissions.");
+    return;
+  }
+
   const articleExists = await existenceVerifier(Article, {
     _id: req.params.id
   });
@@ -110,10 +126,13 @@ router.put("/:id", async (req, res) => {
     res.status(404).send("404 Not Found: Article does not exist.");
   } else {
     const article = articleExists;
+    delete req.body.createdBy;
+    delete req.body.createdAt;
+    req.body.updatedAt = new Date().toISOString();
+    req.body.updatedBy = req.id;
+
     _.merge(article, req.body);
     Object.entries(req.body).forEach(([key]) => article.markModified(key));
-    article.updatedAt = new Date().toISOString();
-    // article.updatedBy = req.id;
 
     article.save(err => {
       if (err) {
@@ -131,7 +150,13 @@ router.put("/:id", async (req, res) => {
  * @param {String} id 删除文章的 ID
  * @returns No Content 或 Not Found
  */
-router.delete("/:id", (req, res) => {
+router.delete("/:id", verifyToken, async (req, res) => {
+  // 只有管理员能够删除文章。
+  if (!(await existenceVerifier(User, { _id: req.id, group: "admin" }))) {
+    res.status(401).send("401 Unauthorized: Insufficient permissions.");
+    return;
+  }
+
   Article.findByIdAndDelete(req.params.id, (err, article) => {
     if (err) {
       res.status(500).send("500 Internal server error.");
